feat : 지하철역 삭제 인수 테스트 만들기 (실습)
